+++ src/components/product/Disclaimer/Disclaimer.stories.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Disclaimer } from './Disclaimer';
+
+export default {
+  title: 'Disclaimer',
+};
+
+export const Default = () => <Disclaimer />;