+++ src/components/product/ServiceStatusInfo/ServiceStatusInfo.tsx
@@ -13,6 +13,7 @@ import {
   OffChainContainer,
   MobileOffChainContainer,
   ExtraContent,
+  DisclaimerLink,
 } from './styles';
 
 const { Text } = Typography;
@@ -35,6 +36,8 @@ type ServiceStatusInfoDetails = {
     setSeconds: (value: number | undefined) => void;
   }) => void;
   onMinimizeToggle?: (isMinimized: boolean) => void;
+  // show the disclaimer
+  showDisclaimer?: boolean;
 };
 
 const timerStyle = { minWidth: '36px' };
@@ -50,6 +53,7 @@ export const ServiceStatusInfo = ({
   extraMd,
   onTimerFinish,
   onMinimizeToggle,
+  showDisclaimer,
 }: ServiceStatusInfoDetails) => {
   const screens = useBreakpoint();
   const canMinimize = !screens.xl;
@@ -174,6 +178,12 @@ export const ServiceStatusInfo = ({
       >
         {canMinimize ? '' : 'Minimize'}
       </Button>
+
+      {showDisclaimer && (
+        <DisclaimerLink href="/disclaimer" rel="noreferrer">
+          Disclaimer
+        </DisclaimerLink>
+      )}
     </ContractsInfoContainer>
   );
 };