+++ HISTORY.md
@@ -44,3 +44,8 @@
 ## 0.3.8
 
 - `sendTransaction` util function added
+
+## 0.3.9
+
+- `ServiceStatusInfo` disclaimer link added
+- `Disclaimer` component added